// src/shared/ui/Button/libs/types.ts
export type TButtonStyles =
  | 'solid'
  | 'outlined'
  | 'ghost'
  | 'soft'
  | 'link'
  | 'white';

export type TButtonSizes = 'small' | 'default' | 'large';
export type TButtonShapes = 'default' | 'pilled';
export type TButtonColors =
  | 'dark'
  | 'gray'
  | 'green'
  | 'blue'
  | 'red'
  | 'yellow'
  | 'light';

export type TButtonFontWeight =
  | 'font-thin'
  | 'font-extralight'
  | 'font-light'
  | 'font-normal'
  | 'font-medium'
  | 'font-semibold'
  | 'font-bold'
  | 'font-extrabold'
  | 'font-black';

export type TButtonFontSize =
  | 'font-text-xs'
  | 'font-text-sm'
  | 'font-text-base'
  | 'font-text-lg'
  | 'font-text-xl'
  | 'font-text-2xl'
  | 'font-text-3xl'
  | 'font-text-4xl'
  | 'font-text-5xl';